Añadidio mensaje de confirmacion para la eliminacion de voluntarios. Cambiada la sentencia SQL de busqueda de Voluntarios (LIKE en lugar de =).

+++ src/Controladores/Voluntario/ControladorVoluntario.java
@@ -457,10 +457,12 @@ public class ControladorVoluntario {
 
         @Override
         public void actionPerformed(ActionEvent ae) {
-            if (eliminarVoluntario(voluntario_temp.getNIF())) {
-                vista.getPanelVoluntarioDatos().setTextLabelError("El Voluntario ha sido eliminado del sistema.");
-            } else {
-                vista.getPanelVoluntarioDatos().setTextLabelError("Error : el Voluntario no ha sido eliminado del sistema.");
+            if(JOptionPane.showConfirmDialog(vista, "¿Seguro que desea eliminar el Voluntario?", "Eliminar Voluntario", JOptionPane.YES_NO_OPTION)==JOptionPane.YES_OPTION){
+                if (eliminarVoluntario(voluntario_temp.getNIF())) {
+                    vista.getPanelVoluntarioDatos().setTextLabelError("El Voluntario ha sido eliminado del sistema.");
+                } else {
+                    vista.getPanelVoluntarioDatos().setTextLabelError("Error : el Voluntario no ha sido eliminado del sistema.");
+                }
             }
             
         }
